review fixes. updated tests. added new tests variants for code.

+++ internal/staticlint/staticio/staticio_test.go
@@ -76,25 +76,64 @@ func TestChecksSA(t *testing.T) {
 }
 
 func TestChecksFromConfig(t *testing.T) {
-	// Create a temporary config file for testing
-	tempConfigPath := createTempConfigFile(t, `{"Staticcheck": ["S1000", "S1001", "ST1000", "ST1005", "QF1003", "QF1010"]}`)
-	defer func() {
-		_ = os.Remove(tempConfigPath)
-	}()
-
-	// Call the function with the temporary config file
-	result, err := ChecksFromConfig(tempConfigPath)
-	if err != nil {
-		t.Fatalf("ChecksFromConfig returned an error: %v", err)
+	type args struct {
+		createTmpFile bool
+		fileData      string
 	}
-
-	// Add your assertions based on the expected result
-	// For example, check the length of the result or specific analyzers
-	if len(result) != 6 {
-		t.Errorf("Expected 6 analyzers, got %d", len(result))
+	tests := []struct {
+		name      string
+		args      args
+		wantCount int
+	}{
+		{
+			name: "valid",
+			args: args{
+				fileData:      `{"Staticcheck": ["S1000", "S1001", "ST1000", "ST1005", "QF1003", "QF1010"]}`,
+				createTmpFile: true,
+			},
+			wantCount: 6,
+		},
+		{
+			name: "valid with incorrect analyzers",
+			args: args{
+				fileData:      `{"Staticcheck": ["S1000", "S1001", "S000", "ST1005", "QF10", "QF1010"]}`,
+				createTmpFile: true,
+			},
+			wantCount: 4,
+		},
+		{
+			name: "failed to read file",
+			args: args{
+				fileData:      `{"Staticcheck": ["S1000", "S1001", "S000", "ST1005", "QF10", "QF1010"]}`,
+				createTmpFile: false,
+			},
+			wantCount: 0,
+		},
 	}
+	for _, ttCommon := range tests {
+		tt := ttCommon
+		t.Run(tt.name, func(t *testing.T) {
+			t.Parallel()
+			// Create a temporary config file for testing
 
-	// Add more assertions as needed
+			tempConfigPath := ""
+			if tt.args.createTmpFile {
+				tempConfigPath = createTempConfigFile(t, tt.args.fileData)
+				defer func() {
+					_ = os.Remove(tempConfigPath)
+				}()
+			}
+
+			result, err := ChecksFromConfig(tempConfigPath)
+			if err != nil && tt.args.createTmpFile {
+				t.Fatalf("ChecksFromConfig returned an error: %v", err)
+			}
+
+			if len(result) != tt.wantCount {
+				t.Errorf("Expected %d analyzers, got %d", tt.wantCount, len(result))
+			}
+		})
+	}
 }
 
 func createTempConfigFile(t *testing.T, content string) string {
@@ -106,7 +145,7 @@ func createTempConfigFile(t *testing.T, content string) string {
 		_ = tmpfile.Close()
 	}()
 
-	if _, err := tmpfile.Write([]byte(content)); err != nil {
+	if _, err = tmpfile.Write([]byte(content)); err != nil {
 		t.Fatalf("Error writing to temp file: %v", err)
 	}
 
@@ -114,16 +153,81 @@ func createTempConfigFile(t *testing.T, content string) string {
 }
 
 func TestReadConfigFromFile(t *testing.T) {
-	tempConfigPath := createTempConfigFile(t, `{"Staticcheck": ["S1000", "S1001", "ST1000", "ST1005", "QF1003", "QF1010"]}`)
-	defer func() {
-		_ = os.Remove(tempConfigPath)
-	}()
-
-	// Call the function with the temporary config file
-	result, err := readConfigFromFile(tempConfigPath)
-	if err != nil {
-		t.Fatalf("readConfigFromFile returned an error: %v", err)
+	type args struct {
+		createTmpFile bool
+		fileData      string
 	}
+	type want struct {
+		wantCount int
+		wantErr   bool
+	}
+	tests := []struct {
+		name string
+		args args
+		want want
+	}{
+		{
+			name: "valid",
+			args: args{
+				fileData:      `{"Staticcheck": ["S1000", "S1001", "ST1000", "ST1005", "QF1003", "QF1010"]}`,
+				createTmpFile: true,
+			},
+			want: want{
+				wantCount: 6,
+			},
+		},
+		{
+			name: "valid with incorrect analyzers",
+			args: args{
+				fileData:      `{"Staticcheck": ["S1000", "S1001", "S000", "ST1005", "QF10", "QF1010"]}`,
+				createTmpFile: true,
+			},
+			want: want{
+				wantCount: 6,
+			},
+		},
+		{
+			name: "failed to read file",
+			args: args{
+				fileData:      `{"Staticcheck": ["S1000", "S1001", "S000", "ST1005", "QF10", "QF1010"]}`,
+				createTmpFile: false,
+			},
+			want: want{
+				wantCount: 6,
+			},
+		},
+		{
+			name: "incorrect json body",
+			args: args{
+				fileData:      `{"Staticcheck": ["S1000", "S1001", "S000", "ST1005", "QF10", "QF1010"`,
+				createTmpFile: true,
+			},
+			want: want{
+				wantErr:   true,
+				wantCount: 6,
+			},
+		},
+	}
+	for _, ttCommon := range tests {
+		tt := ttCommon
+		t.Run(tt.name, func(t *testing.T) {
+			tempConfigPath := ""
+			if tt.args.createTmpFile {
+				tempConfigPath = createTempConfigFile(t, tt.args.fileData)
+				defer func() {
+					_ = os.Remove(tempConfigPath)
+				}()
+			}
 
-	assert.Equal(t, 6, len(result.Staticcheck))
+			// Call the function with the temporary config file
+			result, err := readConfigFromFile(tempConfigPath)
+			if tt.args.createTmpFile && !tt.want.wantErr {
+				if err != nil {
+					t.Fatalf("readConfigFromFile returned an error: %v", err)
+				}
+
+				assert.Equal(t, tt.want.wantCount, len(result.Staticcheck))
+			}
+		})
+	}
 }
